v0.33 #1-6

+++ src/interfaces/Photogenic.java
@@ -7,7 +7,7 @@ public interface Photogenic {
 
 	public String getAvatar() throws SlickException;
 	
-	public void drawScaled(Graphics g, int x, int y, int width, int height);
+	public void drawScaled(Graphics g, int x, int y, float width, float height);
 	
 	public void instantiate() throws SlickException;
